# docs/swagger.yaml
definitions:
  CountStatistics:
    properties:
      long:
        example: 10
        type: integer
      short:
        example: 5
        type: integer
    type: object
  Error:
    properties:
      error:
        type: string
    type: object
  Request:
    properties:
      url:
        type: string
    type: object
  Response:
    properties:
      url:
        type: string
    type: object
  Statistics:
    properties:
      counts:
        $ref: '#/definitions/CountStatistics'
      timings:
        $ref: '#/definitions/TimingStatistics'
    type: object
  TimingStatistics:
    properties:
      long:
        example: "2020-11-10 12:00:05"
        type: string
      short:
        example: "2020-10-23 01:33:45"
        type: string
    type: object
info:
  contact: {}
  title: simple-url-shortener API
  version: "0.1"
paths:
  /long:
    post:
      consumes:
      - application/json
      parameters:
      - description: Short URL
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/Request'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Error'
      summary: Get the origin URL by short URL
  /short:
    post:
      consumes:
      - application/json
      parameters:
      - description: Origin URL
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/Request'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Error'
      summary: Create a new short URL
  /statistics:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Statistics'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Error'
      summary: Getting statistics on URLs
swagger: "2.0"